requirements.txt
pandas
scikit-learn
streamlit
joblib
pyyaml